feat: add consumption and stock resource

+++ src/main/java/com/fawry/store/dtos/StockDto.java
@@ -7,6 +7,6 @@ public record StockDto(
         Long id,
         Long storeId,
         Long productId,
-        Long quantity
+        int quantity
 ) {
 }
